# Changed to a simpler sum.