--- a/sdk/metric/internal/sum_test.go
+++ b/sdk/metric/internal/sum_test.go
@@ -18,8 +18,11 @@
 package internal // import "go.opentelemetry.io/otel/sdk/metric/internal"
 
 import (
+	"context"
+	"fmt"
 	"sync"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -58,7 +61,8 @@ func testSumAggregation[N int64 | float64](t *testing.T, agg Aggregator[N]) {
 
 	extra := make(map[attribute.Set]struct{})
 	got := make(map[attribute.Set]N)
-	for _, a := range agg.flush() {
+	flush := agg.flush()
+	for _, a := range flush {
 		got[*a.Attributes] = a.Value.(SingleValue[N]).Value
 		extra[*a.Attributes] = struct{}{}
 	}
@@ -75,5 +79,78 @@ func testSumAggregation[N int64 | float64](t *testing.T, agg Aggregator[N]) {
 	assert.Lenf(t, extra, 0, "unknown values added: %v", extra)
 }
 
-func TestInt64Sum(t *testing.T)   { testSumAggregation(t, NewSum[int64](NewInt64)) }
-func TestFloat64Sum(t *testing.T) { testSumAggregation(t, NewSum[float64](NewFloat64)) }
+func TestInt64Sum(t *testing.T)   { testSumAggregation(t, NewSum[int64]()) }
+func TestFloat64Sum(t *testing.T) { testSumAggregation(t, NewSum[float64]()) }
+
+func benchmarkSumAggregation[N int64 | float64](b *testing.B, agg Aggregator[N], count int) {
+	attrs := make([]attribute.Set, count)
+	for i := range attrs {
+		attrs[i] = attribute.NewSet(attribute.Int("value", i))
+	}
+
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		for _, attr := range attrs {
+			agg.Aggregate(1, &attr)
+		}
+		agg.flush()
+	}
+}
+
+func BenchmarkInt64Sum(b *testing.B) {
+	for _, n := range []int{10, 50, 100} {
+		b.Run(fmt.Sprintf("count-%d", n), func(b *testing.B) {
+			benchmarkSumAggregation(b, NewSum(NewInt64), n)
+		})
+	}
+}
+func BenchmarkFloat64Sum(b *testing.B) {
+	for _, n := range []int{10, 50, 100} {
+		b.Run(fmt.Sprintf("count-%d", n), func(b *testing.B) {
+			benchmarkSumAggregation(b, NewSum(NewFloat64), n)
+		})
+	}
+}
+
+var aggsStore []Aggregation
+
+// This isn't a perfect benchmark, because we don't get consistant writes. I would probably remove it for production.
+func benchmarkSumAggregationParallel[N int64 | float64](b *testing.B, agg Aggregator[N]) {
+	attrs := make([]attribute.Set, 100)
+	for i := range attrs {
+		attrs[i] = attribute.NewSet(attribute.Int("value", i))
+	}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	b.Cleanup(cancel)
+
+	for i := 0; i < 4; i++ {
+		go func(i int) {
+			for {
+				if ctx.Err() != nil {
+					return
+				}
+				for j := 0; j < 25; j++ {
+					agg.Aggregate(1, &attrs[i*25+j])
+				}
+			}
+		}(i)
+	}
+
+	agg.flush()
+	b.ResetTimer()
+
+	for n := 0; n < b.N; n++ {
+		aggsStore = agg.flush()
+		time.Sleep(time.Microsecond)
+
+	}
+}
+
+func BenchmarkInt64SumParallel(b *testing.B) {
+	benchmarkSumAggregationParallel(b, NewSum(NewInt64))
+}
+func BenchmarkFloat64SumParallel(b *testing.B) {
+	benchmarkSumAggregationParallel(b, NewSum(NewFloat64))
+}